review comments: move back some methods and clean up wording
